[opt] make select-to-bitwise-ops as an independent pass
