+++ xls/passes/BUILD
@@ -97,6 +97,7 @@ xls_pass_registry(
         ":label_recovery_pass",
         ":lut_conversion_pass",
         ":map_inlining_pass",
+        ":select_to_bitwise_ops_pass",
         ":next_value_optimization_pass",
         ":non_synth_separation_pass",
         ":proc_state_array_flattening_pass",
@@ -909,6 +910,34 @@ xls_pass(
         "@abseil-cpp//absl/status",
         "@abseil-cpp//absl/status:statusor",
         "@abseil-cpp//absl/strings",
+        "@abseil-cpp//absl/strings:str_format",
+        "@abseil-cpp//absl/types:span",
+    ],
+)
+
+xls_pass(
+    name = "select_to_bitwise_ops_pass",
+    srcs = ["select_to_bitwise_ops_pass.cc"],
+    hdrs = ["select_to_bitwise_ops_pass.h"],
+    pass_class = "xls::SelectToBitwiseOpsPass",
+    deps = [
+        ":lazy_ternary_query_engine",
+        ":optimization_pass",
+        ":pass_base",
+        ":query_engine",
+        ":stateless_query_engine",
+        ":union_query_engine",
+        "//xls/common/status:status_macros",
+        "//xls/ir",
+        "//xls/ir:bits",
+        "//xls/ir:bits_ops",
+        "//xls/ir:op",
+        "//xls/ir:source_location",
+        "//xls/ir:value",
+        "@abseil-cpp//absl/container:flat_hash_set",
+        "@abseil-cpp//absl/log",
+        "@abseil-cpp//absl/status",
+        "@abseil-cpp//absl/status:statusor",
         "@abseil-cpp//absl/strings:str_format",
         "@abseil-cpp//absl/types:span",
     ],
@@ -3134,6 +3163,28 @@ cc_test(
         "//xls/ir:function_builder",
         "//xls/ir:ir_matcher",
         "//xls/ir:ir_test_base",
+        "@abseil-cpp//absl/status:statusor",
+        "@googletest//:gtest",
+    ],
+)
+
+cc_test(
+    name = "select_to_bitwise_ops_pass_test",
+    srcs = ["select_to_bitwise_ops_pass_test.cc"],
+    deps = [
+        ":select_to_bitwise_ops_pass",
+        ":optimization_pass",
+        ":pass_base",
+        "//xls/common:xls_gunit_main",
+        "//xls/common/status:matchers",
+        "//xls/common/status:status_macros",
+        "//xls/ir",
+        "//xls/ir:bits",
+        "//xls/ir:function_builder",
+        "//xls/ir:ir_matcher",
+        "//xls/ir:ir_test_base",
+        "//xls/solvers:ir_equivalence_testutils",
+        "@abseil-cpp//absl/status:status_matchers",
         "@abseil-cpp//absl/status:statusor",
         "@googletest//:gtest",
     ],
