+++ xls/passes/select_simplification_pass.h
@@ -172,17 +172,6 @@ namespace xls {
 //     potentially more efficient components, and exposing further optimization
 //     opportunities.
 //
-// 16. **Converting Binary `Select` to Masked XOR**:
-//     At higher optimization levels (`SplitsEnabled`), a binary `select` whose
-//     two arms provably differ by a fully-known XOR term (e.g.
-//     `select(p, [x, xor(x, c)])`) is strength-reduced into an XOR with an
-//     AND-masked constant: `xor(x, and(sign_ext(p, W), c))`, with `p`
-//     complemented when the XOR arm is the on-false case. The arms need not
-//     literally form `xor(base, c)`: per-piece xor-with-constant differences
-//     introduced by earlier passes (e.g. `concat_simp`) are also recognized.
-//     This is the select pattern generated by the per-step of an unrolled
-//     CRC/shift loop.
-//
 // The pass is typically run in a fixed-point manner, as each transformation can
 // expose new opportunities for other simplifications. The combination of
 // various analysis techniques (ternary, range, bit provenance) and a wide
