+++ xls/passes/select_to_bitwise_ops_pass.h
@@ -1,0 +1,67 @@
+// Copyright 2026 The XLS Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef XLS_PASSES_SELECT_TO_BITWISE_OPS_PASS_H_
+#define XLS_PASSES_SELECT_TO_BITWISE_OPS_PASS_H_
+
+#include <string_view>
+
+#include "absl/status/statusor.h"
+#include "xls/ir/function_base.h"
+#include "xls/passes/optimization_pass.h"
+#include "xls/passes/pass_base.h"
+
+namespace xls {
+
+// Converts a binary select whose two arms provably differ by a fully-known
+// XOR term (e.g. `sel(p, [x, xor(x, c)])`) into an XOR with an AND-masked
+// constant: `xor(x, and(sign_ext(p, W), c))`, with `p` complemented when the
+// XOR arm is the on-false case. The arms need not literally form
+// `xor(base, c)`: per-piece xor-with-constant differences introduced by
+// earlier passes (e.g. `concat_simp`) are also recognized. This is the select
+// pattern generated by the per-step of an unrolled CRC/shift loop.
+//
+// This pass is separate from select simplification (and runs at the very end
+// of optimization) by design:
+//
+//  * The rewrite is value-based: it only fires when the query engine can prove
+//    the two arms differ by a fully-known constant, unlike the rest of select
+//    simplification which reasons about the shape of the select itself.
+//  * It deliberately consumes shapes that `concat_simp` produces (wide xor
+//    limbs split into per-slice concats). `concat_simp` runs inside the
+//    fixed-point "simp" compound, and running this rewrite in the same
+//    fixed point would leave the two transforms repeatedly reshaping the same
+//    IR instead of converging.
+//  * It converts the select into an `and`-masked xor, which value-based
+//    analyses reason through less easily, so it is deferred until after those
+//    analyses (e.g. conditional specialization) have already run.
+//
+// Runs only at optimization levels where split-based rewrites are enabled
+// (SplitsEnabled).
+class SelectToBitwiseOpsPass : public OptimizationFunctionBasePass {
+ public:
+  static constexpr std::string_view kName = "select_to_bitwise_ops";
+  SelectToBitwiseOpsPass()
+      : OptimizationFunctionBasePass(kName, "Select to Bitwise Ops") {}
+  ~SelectToBitwiseOpsPass() override = default;
+
+ protected:
+  absl::StatusOr<bool> RunOnFunctionBaseInternal(
+      FunctionBase* f, const OptimizationPassOptions& options,
+      PassResults* results, OptimizationContext& context) const override;
+};
+
+}  // namespace xls
+
+#endif  // XLS_PASSES_SELECT_TO_BITWISE_OPS_PASS_H_
